+++ internal/digest/generator_test.go
@@ -1,0 +1,14 @@
+package digest
+
+import (
+	"fmt"
+	"testing"
+)
+
+func TestDigest(t *testing.T) {
+	d, err := Generate("admin", "adminPw")
+	if err != nil {
+		t.Error(err)
+	}
+	fmt.Println(d)
+}
